Major push towards 1.13 support NOTE: This will NOT compile as is. You can manage to get it to compile with the removal of a few of the volatile code handler classes depending on what version you want it to run on - will be fixed at a later date.
The reason behind why this will not properly compile is due to how the build system is currently set up. I will be revamping this in the next few commits and hopefully allowing for a version that fully supports all the way from 1.8 to 1.13.

Note for Red/Finder: Revisit the TODO's in the commits from either me or Sam/the MagicSpells dev team.